use 'i18n.resolvedLanguage', not 'language'
This was added in a relatively recent version of i18next; some online resources haven't updated to reflect it
https://www.i18next.com/overview/api#resolvedlanguage

`language` is what i18next detected or we have set it to. It could be either a locale code like 'en-US' or a plain language like 'en'.

`resolvedLanguage` is what language is actually being used for translation, after considering what is detected, what langauges we provided, and what fallbacks we specified.

In our case this should always be either 'en', 'es', or 'lo' - the 3 languages we support at this time.

So `resolvedLanguage` is the safe property to use and we should avoid `language`.