for conditional survey eval, spread trip props
GIven that the configs are JSON and we can only give the "showsIf" conditions as strings, it will be annoying if they get too lengthy.
This will allow us to reference trip properties directly (e.g. "end_loc" instead of "tlEntry.end_loc") which is a bit more concise and readable.