EnketoModal: allow appConfig to be initially null
It is actually expected for appConfig to be initially null here, and we don't need to show an error message for it.
EnketoModal is not rendered until it is needed; it hasn't called useAppConfig() in advance. So until useAppConfig() resolves, appConfig is null.
Once it does resolve and appConfig is set, this useEffect runs again and initSurvey() gets called.