add AlertManager for global alerts
There's a bunch of places in the app where we need to show a SnackBar or "toast"-style message.
But to do this, we've been having to keep each SnackBar's visibility as state so it can show/hide depeding on its "visible" value.

It would be simpler to be able to do this from anywhere in the codebase, though, not just by setting local state.
So I refactored AlertBar to behave as a global stack of alert messages. AlertBar is instantiated once at the top level in App.tsx. A static class called AlertManager can be used to add messages to this stack from anywhere in the codebase.
As a result, components no longer need to have their own instance of AlertBar and they don't have to control the visibility themselves.
They just call AlertManager.addMessage when they want to show a message.