make Log Out button a NavBarButton; tweak NavBarButton
On the other screens of the UI, we use a NavBarButton for clickable buttons with text. I realized the Log Out button is inconsistent with this, so I changed it to a NavBarButton.
I tested this with a screen reader and it is still accessible (since it uses the Button component it is grouped as and recognized as 1 element with role 'button')

Using a NavBarButton made the log out icon smaller than before, so while making this change I revised the NavBarButton props. I added an optional "iconSize" that can be used to override the default 20.
I removed "onPressAction" because the Button component already has "onPress" which we can just use.
I organized the styles for both NavBar and NavBarButton, and made them able to be extended on by props.