fix uploadService arrow functions 'this'
in 31d472e85814f898236972c9aa249c3c7996b597, the reader.onprogress, reader.onerror, and reader.onload functions here were changed from regular `function` declarations to arrow functions.

Usually this makes no difference in the operation of the code, but there is one exception.
Arrow functions don't get their own `this`, they just inherit it from the scope. When I swapped this out and made it an arrow function, I didn't notice it used `this`.

In this instance, `this` was the same thing as `reader`, which we already have a reference to. So it's an easy swap.