downscale tiles on TripCard LeafletView
Using lower res images for the tiles has a small performance benefit with scrolling.
Since these are small maps, they do not need to be high-res.
In fact, the text on the maps is larger and more readable with the lower res tiles, so this works out anyway.

Note: The maps on LabelDetailsScreen still use high-res tiles because downscaleTiles is not set true there.