only show Leaflet + OSM attribution on the first trip
Every trip on the Label screen has a map, and we've been showing the Leaflet and OSM attribution on every single one.
This makes the DOM a little bit cluttered. Trimming it down could help performance marginally.
Also, it's distracting and unnecessary to have it on every map.

Leaflet does not care if you disable the attribution. But we will still attribute Leaflet because it's nice to do so :)

The tiles from OSM are copyrighted; they DO care about attribution and have specific guidelines.
In the guidelines they say "...if multiple static images appear on the same document, one instance of attribution is sufficient."
https://osmfoundation.org/wiki/Licence/Attribution_Guidelines#Static_images

These are static maps so it's good enough to show it on just one map. I think it's fair to put it on the most recent trip since that one will appear first in the list. This way, attribution will show on the first render but will not be distracting when scrolling up to other trips.

Note that the full attribution always appears on the map on the LabelDetailsScreen, since that one is interactive and it's the only one on the page. So even on static maps where the attribution has been hidden, the attribution is 1 click away.