set innerHTML directly instead of using html-react-parser
We don't need this library because we can set innerHTML directly using dangerouslySetInnerHTML.

This would be dangerous if we did not have control over where the html string came from. But we do have control; it is just a cached copy of something that was already rendered.
So it is fine to use in this context.

Also, the actual instance of the Leaflet map can be removed once its HTML is cached.