LeafletView: make the attribution hyperlinks functional
To comply with OSM's attribution guidelines we need to link to the copyright page on their website.

We do include links, but they haven't been working because <a> elements don't work in our Cordova app like they would on the web.
To open a URL we have to use cordova inappbrowser like this.

I did the same for the Leaflet attribution as well.