LeafletView: only cache for TripCard maps
The TripCard maps are being cached because we want to imporve performance on the trips list.

But the maps on the label details screen should still be fully interactive; they should not be cached.

Also, we can append '-downscaled' to the IDs of maps that use downscaled tiles; making them distinct from non-downscaled maps. this way Leaflet doesn't try to mix high-res and low-res tiles on the same map