use reversed FlatList without CSS transforms
https://github.com/e-mission/e-mission-phone/pull/1132
We are trying to improve performance of the timeline scroll.
The way that FlashList / FlatList 'inverted' property works has shown to be a bottleneck. It "uses CSS transforms to flip the entire list and then flip each list item back, which is a performance hit and causes scrolling to be choppy, especially on old iPhones."

An alternative I found is setting the 'flex-direction' to 'column-reverse'. But this only works for FlatList.
'column-reverse' is applied to the content so the items in the list display bottom-to-top, and also applied to the wrapper so the scroll position initializes from the bottom.
This yields better performance!

But there is another issue: on iOS the list flashes for a second while loaded and then goes blank. Only once the user interacts is the list visible.

I think this is an issue with iOS Safari. I came up with a workaround to add a temporary 1px margin to trigger a layout update. It's not ideal, but it's minimally intrusive and won't cause further problems.

Eventually if we fully migrate to React Native, we will not have to worry about Safari at all; we should be able to just use FlatList or FlashList as-is.