move leaflet cache to a stateful hook
There was an issue where map elements were sometimes showing blank when they should have been cached. This is because they were stored in a Map object, in LeafletView.tsx but outside of the component.
It was not kept as state so React didn't know to re-render if it changed.

We can keep this Map object as state in a custom hook that all LeafletViews can use.