refactor TimelineContext to use dateRange
The way this has been working up to now is that there are specific functions to load one week of data - either a specific week or the previous/next week. These functions load new trips and then update the queriedRange after the fact.

But we need this to now work on the Metrics tab as well (where you can select *any* date range) so we are going to need to make some changes.

We need to keep the chosen dates (from datepicker) centrally as state. Then we can watch it for changes; when it changes, we can compare the old and new ranges to see how they intersect: it will be either i) we are loading more data into the past ii) we are loading more data into the future or iii) this is an entirely new range
Once we have determined that, we can load in the new part(s) of the range and handle the new data in the same way as before.

I chose to store the date range as yyyy-mm-dd strings. The queried date range will also be stored as yyyy-mm-dd strings

The initial date range is now set as the default value of `dateRange` in TimelineContext. So the initialy query for the past week can be removed from LabelTab.