unify DateSelect for Label tab and Dashboard tab
The previous commit implemented a central state for the selected timeline range.
The DateSelect component needs to control this, and both the label tab and dashboard tab needs to use it.

DateSelect can be made generic such that it supports either a 'single' selection (which selects a whole week on the Label tab), or it supports a 'range' selection (as on the Dashboard tab).

The DateSelect will display the queried date range, and it will accept an onChoose function that controls what the selection of date or date range will do.

On LabelListScreen and MetricsTab are where DateSelect will be used.
On LabelListScreen, the onChoose function calls loadSpecificWeek from TimelineContext (which sets the range to the week of [3 days before X, 3 days after X].
On MetricsTab, onChoose just sets the TimelineContext date range to the same date range from the datepicker.

MetricsDateSelect is no longer needed.

TimelineScrollList updated to use queriedDateRange instead of queriedRange; also access timeline variables from TimelineContext instead of passing a bunch of props.