revert Enketo buttons from `<Button>`s back to plain HTML
Enketo requires there to be `<a>` elements with specific IDs for the "next" and "save" buttons.
We had nested `<Button>`s inside these to make them match the `<Button>`s we use elsewhere in the UI.
But in some webviews, the touch events would not bubble up from the `<Button>` to the `<a>` elements, causing the buttons to be unclickable except around the margins.

Reverting this back to just use plain `<a>` elements to resolve the bug.
Maybe revisit `<Button>` later because it was nicer looking.