allow missing days for segmentDaysByWeeks
With the server calls, the metrics data always had 1 entry for every date in the query. Even if there was no data for that date, it would give a blank entry.
e-mission-common isn't set up to work like this; it only gives entries on days that had trips.
This change makes segmentDaysByWeeks work if there are missing dates. it requires an extra parameter for 'end date'

I also noticed there is a lot of repeated code between these "Card" components; this should be refactored later