fix ActiveMinutes components
WeeklyActiveMinutesCard.tsx:
Add records even if the total is 0 so that both "Walk" and "Regular bike" will show up in the legend.

DailyActiveMinutesCard.tsx:
The line chart was not showing any points at all. Turns out it's because the timestamp values were strings and needed to be numbers.
We will now add records with 'null' values even if there is no value for a mode on a day. Both "Walk" and "Regular bike" now show in the legend.
Adding 'spanGaps' to the chart options (in Chart.tsx), causes a line on the chart to connect only values that were on consecutive days; otherwise there will be a gap between entries on non-consecutive days.
This gives users a nice way to visualize having a "streak" of days they used a particular active mode.