update types relating to `UnprocessedTrip`s
The main thing I was doing in this commit was adding 'sections' to the type signature of `UnprocessedTrip`. But I also noticed some other things amiss.

`UnprocessedTrip` was missing some other properties; end_ts and start_fmt_time

"LocationCoord" is not needed as it's the same as `Point` from 'geojson'.

`SectionData` was missing a bunch of properties. Once those are filled in, the 'sections' property in `CompositeTrip` and `UnprocessedTrip` can be typed as `SectionData[]`