flesh out more types
Filled in more trip object types, and specifically some literal types that define unprocessed trips. (like "source" will always be 'unprocessed').

CompTripLocations changed from `number[]` for coordinates to geojson's `Position`, since that is more descriptive. Renamed CompTripLocations to CompositeTripLocation (this type represents only 1 location).

Used the CompositeTripLocation type in timelineHelper.
in locations2GeojsonTrajectory, the return type needed `properties` for it to be considered a Geojson `Feature`.

formattedSectionProperties types as the return type of the function