show metrics tab if phone_dashboard_ui is set in app config
Up to now, the dashboard has only been shown for MULTILABEL configurations; on ENKETO configurations it was completely hidden.

We are making the dashboard more configurable -
https://github.com/e-mission/e-mission-docs/issues/1055.
The presence of a new field `metrics`.`phone_dashboard_ui` being defined will cause the dashboard to be shown.
If `phone_dashboard_ui` is not defined, it will fall back to the current behavior (which is to only show dashboard for MULTILABEL configurations).