🔥 Remove code to build JSON objects
In an earlier version of this PR, before I added mocking to the plugin,
we saved JSON objects directly. We created the objects to save by
removing results from the in-memory objects using a destructuring
assignment.

However, after the mocking was added, we create objects in native code.
So we don't need to create them here and can remove these unused
variables.
https://github.com/e-mission/e-mission-phone/pull/1144#discussion_r1564981905