add "confirmedMode" to derived properties & use for conditional surveys
For the fermata project, the survey shown for a trip should depend on what vehicle was detected from bluetooth.
For unprocessed trips, this information is not going to be in the raw trip object. That's why we have "timelineBleMap", which is kept separate from the raw trip objects.
But to access this while evaluating the conditional survey "showsIf" expressions, we'll need to include more things in the eval's scope.

I decided that it would be appropriate to use our "derived properties" hook for this. we can add a new property to derived properties, then we can include all the derived properties in the scope in which the "showsIf" expressions are evaluated.

This will probably be adjusted later.

Note: I also simplified the type definition of DerivedProperties. It now just uses the return type of the useDerivedProperties hook.