only use "ranging" scans for BLE matching
When matching we basically are trying to find the major:minor pair that was sensed the most frequently during the trip.
But should only consider RANGE_UPDATE events for this. The REGION_ENTER and REGION_EXIT events do not have major and minor defined, so we should exclude them for the purpose of matching.

Add a filter condition for this. Also updated function + variable names to make it clearer that we are only considering 'ranging' scans.

Note that before processing, the value of eventType is a string ('RANGE_UPDATE'). But the server changes this to an enum value where 'RANGE_UPDATE' -> 2. So we have to check for both.