cacheResourcesFromConfig: reload the cache when config is downloaded
We recently added the ability to refresh the config from inside the app.
But whenever the config is downloaded for the first time, we cache the resources referenced in it by URL. Subsequent config downloads would still use the resources from the first time.

fetchUrlCached now accepts options to pass through to the fetch API; if cache is 'reload', we will skip checking our localStorage cache for a previously stored value. This option will also cause the fetch API will also skip its own internal cache

The result of this is that when we refresh the config, URL-referenced resources inside it will also be refreshed.