use latest minor version of React 18 (18.3.0)
We were using react 18.2.*, while also using the latest react testing library, which asked for react 18.3, which caused an issue with npm installation.

When React 19 was released recently, they put out 18.3 at the same time (https://react.dev/blog/2024/04/25/react-19-upgrade-guide#react-18-3). 18.3 is the same as 18.2 but with a few deprecation warnings so it's easier to upgrade to React 19.
We're not ready for the React 19 upgrade, but we can use 18.3 now, view those warnings and think about upgrading to 19 later.