🚑️ 🐛 Ensure that we do show the replaced mode when the `mode_of_interest` matches
At some point, we changed the mode and purpose user inputs as objects that were
stored to be full objects, with start and end timestamps, instead of just the
labels. We changed all uses of the MODE and PURPOSE to match it, but apparently
forgot to change this location, where the replaced mode button is conditionally
displayed.

This is a quick change to make the usage here consistent with the rest of the
code so that we can push this out ASAP.

@JGreenlee, please let me know if there is a more principled fix, it is late
and I don't want to experiment any further.

Testing done:
Please see PR