💩 Turn off the background checks for fleet deployments
Since the fleet deployments are installed on work phones, and the background restriction is apparently grayed out on android work phones.
https://github.com/e-mission/e-mission-docs/issues/1070

This is a hack (hence the 💩 emoji) but it will allow us to make progress
through the onboarding, get the logs and then resolve the issue the right way.
Or if we can't figure out how to access it the right way, this can become a
config option for deployments that plan to use work phones.

This should unblock
https://github.com/e-mission/e-mission-docs/issues/1070#issuecomment-2094354832