✅ 🐛 Improve user input matching mocks and use them
The previous unit tests for the input matching assumed that the user input
would be of the form of the label options, so reused the label options as mock
options. However, they actually are of the form of user input objects with data
and metadata and the input as a label.

We create new mock objects with the correct format, and use them in the tests.

This is the reason why https://github.com/e-mission/e-mission-phone/pull/1150
was not caught for so long.

And when we fixed the code, the test broke.
https://github.com/e-mission/e-mission-phone/pull/1150#issuecomment-2094892441
https://github.com/e-mission/e-mission-phone/pull/1150#issuecomment-2094895504

Testing done:

```
npx jest www/__tests__/confirmHelper.test.ts

Test Suites: 1 passed, 1 total
Tests:       11 passed, 11 total
```