display response_count MetricsCard properly
Instead of specifically listing a MetricsCard for each metric, we can read metricsList and map each key to a MetricsCard.

But the structure for response_count is an object of {responded: number, not_responded: number}, while the other metrics are just numbers.
So we need to adjust change the way days get summed up to account for both.