use real data for SurveyComparisonCard
Instead of dummy data, we can now use real values via the "response_count" metric.
Since our data is chunked by day and this comparison is across all surveys for all days loaded, we need a function (getResponsePctForDays) to sum up the 'responded' and 'not_responded' and calculate a percentage. This is called on userMetrics for "you" and on aggMetrics for "others".

It was easy to swap in the real percentages to the existing doughnut charts and they look great!