fix filters issue on 'additions' configs
We skip filters if places are shown in the timeline, so filterInputs stays as the initial value, which is [].
So when we checked if filterInputs has length (line 64), this was always false on configs where places are shown. thus, displayedEntries never got set.

We had no way to distinguish between "haven't done filters initialization yet" and "did filters initialization and there are none".
So I changed the initial value to null. If filterInputs is null, it's awaiting init. If it's [], it's been initialized but there are no filters; we should proceed with displayedEntries.