on config refresh, wait for resources to finish caching
When downloading a config, we look for resources referenced by URL and cache them. By default we don't 'await' this - we just let it happen. This works well for the onboarding flow where the config is loaded and we can fetch resouces to cache in the background while the user is proceeding through onboarding.

But when refreshing a config, the app gets reloaded after the new config is loaded. So if we hadn't finished caching resources when the reload was triggered, those resources didn't get cached. It's not the end of the world, but the first time the user accesses that resource it will be slower.

So now, before reloading we will wait for a promise which tracks whether all of the resources have been cached.
The onboarding flow stays the same because it does doesn't wait for that promise.