refactor metrics "units" & "format" fns to support response_count charts
MetricsCard had been receiving these 2 props: axisUnits and unitFormatFn.
The value of the metric was always a number, so we could just use that for the chart, and for "display" we'd just append the axisUnits as a suffix to the number value.

With response_count, the value of the metric is an object which has counts for "responded" and "not_responded".
For "display", we want the # responded / total responses, e.g. "3/5 responses".
But for the chart we need a single number so it will be the # responded (e.g. 3)

So to deal with units we basically need 3 things: a "convert" fn, a "display" fn, and the units suffix (for the axis of the chart). I call these "unit utils" and I put them in metricsHelper to keep MetricsTab and MetricsCard from being too cluttered.