unify datepickers, loadSpecificWeek -> loadDateRange; simplify
We are unifying the datepickers of the label tab and dashboard tab to both be "range" datepickers (https://github.com/e-mission/e-mission-phone/pull/1138/files#r1606146462)
So there is no need for a function like "loadSpecificWeek" that takes one day as input.
I refactored this into "loadDateRange" which accepts a start and end date, and clamps it within [pipeline start, today].
Then I realized loadMoreDays could be simplified and it can just call loadDateRange once it determines what the range should be.

'setDateRange' no longer needs to be exposed because 'loadDateRange' fulfills that purpose and also handles clamping