improve MetricsTab loading
Breaks the useEffect into 3 smaller chunks.

`readyToLoad` handles the fetching of more days if < 14. Also added a return type to `loadMoreDays` to handle the case where we don't have 14 days, but we can't go back any further; `loadMoreDays` will return false and `readyToLoad` will become true.

Then 2 separate useEffects for `userMetrics` and `aggMetrics`. `userMetrics` depend on readyToLoad, appConfig, timelineIsLoading, timelineMap, and timlineLabelMap, and compute if timelineIsLoading is falsy and all the rest are truthy.
`aggMetrics` depend on readyToLoad, appConfig, and dateRange and get fetched if readyToLoad and appConfig are defined.

The result of this is that `aggMetrics` do not have to wait for the timeline to finish loading; fetching aggMetrics begins once the dateRange is determined to be adequate. Also, `aggMetrics` will not be unnecessarily recomputed if `timelineLabelMap` changes (ie the user labels a trip)