fix timeline initial date range regression
When refactoring into TimelineContext, there was a regression regarding the date range that is initially loaded.
https://github.com/e-mission/e-mission-phone/pull/1142#issuecomment-2141345034

It used to load 7 days before the pipelineEnd up to today. After the change, it would load 7 days before today up to today.
If the user has travel everyday this ends up being equivalent. But if there isn't recent travel, this becomes a problem.

So in TimelineContext, the initial dateRange needs to depend on pipelineRange. instead of giving an initial value upfront, we'll let dateRange be null until pipelineRange is set, at which point we'll also set dateRange.
This means we have to consider some extra cases where dateRange can be null.