prevent concurrent retrievals of dynamic config
While investigating excessive log statements, I found several repetitive logs for config retrieval.
There were several retrievals occuring in parallel because the retrieval is not synchronous but was being requested by multiple parts of the codebase in succession.
A better pattern is to cache the promise rather than only caching the result once the promise completes

I also updated the return type of getConfig to properly reflect that the promise could resolve as `null`.
Throughout the codebase we already account for the possibility of it being null almost everywhere except tests. enketoHelper is the one place I saw that config being null could have been an issue.
For tests, we just tell TypeScript to assert non-null with `!`