🔥 Remove the pinned npm version; use the node default
We had a pinned npm version, but we didn't actually use it,
using the default version in node instead. As long as we have
a reasonbly recent version of node, that seems like it will work.
It is still pinned (just pinned implicitly by the node pin), and we don't
actually use the pinned version now anyway.

Yay for simplicity and not keeping bitrotted code around...