♻️  Use the correct package name for ibeacon plugin
This ensures that we don't "discover" the plugin and try to install it
https://github.com/e-mission/e-mission-phone/pull/1168#issuecomment-2313327905

Without this change:
```
Discovered plugin "cordova-plugin-ibeacon". Adding it to the project
Plugin "com.unarin.cordova.beacon" already installed on android.
Plugin "com.unarin.cordova.beacon" already installed on ios.
```

With this change: those logs don't show up