simplify startprefs
Instead of a fetch() call to get the startup config, let's load the JSON file using ES import syntax. (Rename the file to startupConfig.default.json in the process, get rid of the .sample)
While here, we can tidy this code with async/await instead of promises.
I also noticed that this function doesn't return the consent status; returning it simplifies every place where this is invoked.