✨ Add Expo to the project
Switches the project from React Native web project using Webpack to an Expo web project using Metro bundler

We will now use Metro, the standard bundler for React Native / Expo projects, instead of webpack. Metro is less flexible but more tailored to the use case of a react native / expo project, and switching to it will make our lives easier for future native code migration.

It also allows us to add a small compatibility layer (nativePlugins.ts) which will allow us to run a minimal version of the app in a browser with our native code mocks. If nothing else, this will be useful as a dev tool

Cordova assumes that index.html and its required resources will exist in www/.
By default, Expo export generates the index.html and bundled resources and outputs to dist/.
We can change the output dir to www/, but we should keep our source files separate, so I have moved them to src/.
So www/ now functions as an "output" or "dist" directory to which Metro exports, and from which Cordova builds.
public/index.html is the "template" that the Expo export uses for www/index.html which ensures that the output index.html actually has the cordova.js included

We will no longer have a "serve" configuration, so the cordovabuild config.xml is now the de facto config.xml.